Main auction logic finished

+++ src/main/java/com/codecool/old_book_auction/model/Bid.java
@@ -3,15 +3,15 @@ package com.codecool.old_book_auction.model;
 public class Bid {
     private int id;
     private Bidder bidder;
-    private double bidPrice;
+    private int bidPrice;
 
-    public Bid (int id, Bidder bidder, double bidPrice) {
+    public Bid (int id, Bidder bidder, int bidPrice) {
         this.id = id;
         this.bidder = bidder;
         this.bidPrice = bidPrice;
     }
 
-    public double getBidPrice() {
+    public int getBidPrice() {
         return bidPrice;
     }
 
@@ -23,7 +23,7 @@ public class Bid {
         this.bidder = bidder;
     }
 
-    public void setBidPrice(double bidPrice) {
+    public void setBidPrice(int bidPrice) {
         this.bidPrice = bidPrice;
     }
 }
